Update Dockerfile and docker-compose.yml
This commit updates the Dockerfile to use yarn instead of npm for installing project dependencies and adds the yarn.lock file to the working directory. It also updates the docker-compose.yml file to use the latest version of the mongo image and adds a new volume for the app data.